refactor: move EcallCheck to own file
