+++ riscv_analysis/src/lints/checks.rs
@@ -153,19 +153,6 @@ impl LintPass for DeadValueCheck {
     }
 }
 
-// Check if every ecall has a known call number
-// Check if there are any instructions after an ecall to terminate the program
-pub struct EcallCheck;
-impl LintPass for EcallCheck {
-    fn run(cfg: &Cfg, errors: &mut DiagnosticManager) {
-        for node in cfg {
-            if node.is_ecall() && node.known_ecall().is_none() {
-                errors.push(LintError::UnknownEcall(node.node().clone()));
-            }
-        }
-    }
-}
-
 // TODO deprecate
 // Check if there are any in values to the start of functions that are not args or saved registers
 // Check if there are any in values at the start of a program
